Address @gaschler's comments.
