+++ cartographer_grpc/framework/rpc_handler.h
@@ -17,6 +17,8 @@
 #ifndef CARTOGRAPHER_GRPC_FRAMEWORK_RPC_HANDLER_H
 #define CARTOGRAPHER_GRPC_FRAMEWORK_RPC_HANDLER_H
 
+#include "cartographer_grpc/framework/rpc.h"
+#include "cartographer_grpc/framework/rpc_handler_interface.h"
 #include "cartographer_grpc/framework/type_traits.h"
 #include "glog/logging.h"
 #include "google/protobuf/message.h"
@@ -24,29 +26,6 @@
 
 namespace cartographer_grpc {
 namespace framework {
-
-class Rpc;
-class RpcHandlerInterface {
- public:
-  virtual ~RpcHandlerInterface() = default;
-  virtual void SetRpc(Rpc* rpc) = 0;
-  virtual void OnRequestInternal(
-      const ::google::protobuf::Message* request) = 0;
-  virtual void OnReadsDone() = 0;
-};
-
-using RpcHandlerFactory =
-    std::function<std::unique_ptr<RpcHandlerInterface>(Rpc*)>;
-
-struct RpcHandlerInfo {
-  const google::protobuf::Descriptor* request_descriptor;
-  const google::protobuf::Descriptor* response_descriptor;
-  const RpcHandlerFactory rpc_handler_factory;
-  const grpc::internal::RpcMethod::RpcType rpc_type;
-};
-
-void SendInternal(std::unique_ptr<::google::protobuf::Message> message,
-                  Rpc* rpc);
 
 template <typename Incoming, typename Outgoing>
 class RpcHandler : public RpcHandlerInterface {
@@ -59,11 +38,11 @@ class RpcHandler : public RpcHandlerInterface {
   void SetRpc(Rpc* rpc) override { rpc_ = rpc; }
   void OnRequestInternal(const ::google::protobuf::Message* request) override {
     DCHECK(dynamic_cast<const RequestType*>(request));
-    OnRequest(static_cast<const RequestType*>(request));
+    OnRequest(static_cast<const RequestType&>(*request));
   }
-  virtual void OnRequest(const RequestType* request) = 0;
+  virtual void OnRequest(const RequestType& request) = 0;
   void Send(std::unique_ptr<ResponseType> response) {
-    SendInternal(std::move(response), rpc_);
+    rpc_->Write(std::move(response));
   }
 
  private:
